unify removal of services

+++ packages/google-compute-engine/packaging/google-compute-engine.spec
@@ -56,15 +56,20 @@ cp -a src/lib/udev/rules.d/* %{buildroot}/%{_udevrulesdir}
 %config /etc/rsyslog.d/*
 %config /etc/sysctl.d/*
 
-%post
-# Remove old services.
-for svc in google-ip-forwarding-daemon google-network-setup \
-  google-network-daemon google-accounts-daemon google-clock-skew-daemon \
-  google-instance-setup; do
-    if systemctl is-enabled ${svc}.service >/dev/null 2>&1; then
-      systemctl disable ${svc}.service >/dev/null 2>&1 || :
-      if [ -d /run/systemd/system ]; then
-        systemctl stop ${svc}.service >/dev/null 2>&1 || :
+%pre
+if [ $1 -gt 1 ] ; then
+  # This is an upgrade. Stop and disable services previously owned by this
+  # package, if any.
+  for svc in google-ip-forwarding-daemon google-network-setup \
+    google-network-daemon google-accounts-daemon google-clock-skew-daemon \
+    google-instance-setup; do
+      if systemctl is-enabled ${svc}.service >/dev/null 2>&1; then
+        systemctl --no-reload disable ${svc}.service >/dev/null 2>&1 || :
+        if [ -d /run/systemd/system ]; then
+          systemctl stop ${svc}.service >/dev/null 2>&1 || :
+        fi
       fi
-    fi
-done
+  done
+  systemctl daemon-reload >/dev/null 2>&1 || :
+fi
+
